Refactoring reducers and actions #57
