+++ src/render/Reducers/HostPeerReducer.ts
@@ -1,0 +1,61 @@
+import { IHostPeerState } from "../redux/State";
+import { Action, ActionType } from "../Actions/Action";
+import { HostPeerAction, IAddSignalDataAction, IClearSignalDataAction, ICreatePeerAction, ISetPeerStatusAction } from "../Actions/HostPeerActions";
+import { addSignalData } from "../utils/ReduxUtils";
+import { DataType } from "../utils/Definitions";
+
+const initialHostPeerState: IHostPeerState = {
+    hostPeers: [],
+};
+
+const hostPeerState = (state: IHostPeerState = initialHostPeerState, action: Action<HostPeerAction>) => {
+    const types = ActionType.hostPeerAction;
+
+    switch (action.type) {
+        case types.addClientSignalDataAction:
+            return addSignalData(state, action as IAddSignalDataAction, DataType.CLIENT);
+
+        case types.addHostSignalDataAction:
+            return addSignalData(state, action as IAddSignalDataAction, DataType.HOST);
+
+        case types.clearSignalDataAction:
+            return Object.assign({}, state, {
+                hostPeers: state.hostPeers.map((peer) => {
+                    if (peer.clientID === (action as IClearSignalDataAction).id) {
+                        Object.assign(peer, {
+                            hostSignalData: [],
+                        });
+                    }
+                    else {
+                        return peer;
+                    }
+                }),
+            });
+
+        case types.createPeerAction:
+            return Object.assign({}, state, {
+                hostPeers: state.hostPeers.concat({
+                    peerStatus: false,
+                    clientID: (action as ICreatePeerAction).clientID,
+                    clientSignalData: (action as ICreatePeerAction).signalData,
+                    hostSignalData: [],
+                }),
+            });
+
+        case types.setPeerStatusAction:
+            return Object.assign({}, state, {
+                hostPeers: state.hostPeers.map((peer) => {
+                    if (peer.clientID === (action as ISetPeerStatusAction).clientID) {
+                        return Object.assign(peer, {
+                            peerStatus: (action as ISetPeerStatusAction).status,
+                        });
+                    }
+                }),
+            });
+
+        default:
+            return state;
+    }
+};
+
+export default hostPeerState;
