+++ src/render/Reducers/WindowReducer.ts
@@ -1,0 +1,26 @@
+import { IWindowState } from "../redux/State";
+import { Action, ActionType } from "../Actions/Action";
+import { WindowAction, IResizePageAction } from "../Actions/WindowActions";
+
+const initialWindowState: IWindowState = {
+    height: 600,
+    width: 800,
+};
+
+const windowState = (state: IWindowState = initialWindowState, action: Action<WindowAction>): IWindowState => {
+    const types = ActionType.windowAction;
+
+    switch (action.type) {
+
+        case types.resizePageAction:
+            return Object.assign({}, state, {
+                height: (action as IResizePageAction).height,
+                width: (action as IResizePageAction).width,
+            });
+
+        default:
+            return state;
+    }
+};
+
+export default windowState;
